# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Forgot password",
                "operationId": "37929c5357e2918c3d1d0985545011ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset link sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "responseStatus": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "responseMessage": {
                                            "type": "string",
                                            "example": "Lien envoyé!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "responseStatus": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "responseMessage": {
                                            "type": "string",
                                            "example": "Lien non envoyé!"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Please wait before retrying."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "responseStatus": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "responseMessage": {
                                            "type": "string",
                                            "example": "Veuillez vérifier l'adresse e-mail!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "responseStatus": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "responseMessage": {
                                            "type": "string",
                                            "example": "Une erreur est survenue!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset password",
                "operationId": "8128315ea5ef2b62b30126390e8043f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "token_here"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new_password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "new_password_confirmation"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "responseStatus": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "responseMessage": {
                                            "type": "string",
                                            "example": "Mot de passe réinitialisé avec succès!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "responseStatus": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "responseMessage": {
                                            "type": "string",
                                            "example": "Veuillez vérifier le mot de passe!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The password field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "responseStatus": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "responseMessage": {
                                            "type": "string",
                                            "example": "Veuillez vérifier le mot de passe!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "password": [
                                                    "The password confirmation does not match."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "responseCode": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "responseStatus": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "responseMessage": {
                                            "type": "string",
                                            "example": "Une erreur est survenue!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The password field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}